@@ -19,14 +19,23 @@
     "@types/supertest": "^2.0.9",
     "@zeit/ncc": "^0.21.1",
     "jest": "^25.1.0",
-    "nodemon": "^2.0.2",
     "sqlite3": "^4.1.1",
     "supertest": "^4.0.2",
     "ts-jest": "^26.1.0",
-    "ts-node": "^8.10.2",
-    "typescript": "^3.9.5"
+    "@typescript-eslint/eslint-plugin": "^4.14.0",
+    "@typescript-eslint/parser": "^4.14.0",
+    "eslint": "^7.18.0",
+    "eslint-config-prettier": "^7.2.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-prettier": "^3.3.1",
+    "nodemon": "^2.0.7",
+    "prettier": "^2.2.1",
+    "ts-node": "^9.1.1",
+    "typescript": "^4.1.3"
   },
   "dependencies": {
+    "io-ts": "^2.2.14",
+    "fp-ts": "^2.9.5",
     "@koa/cors": "^3.0.0",
     "bcryptjs": "^2.4.3",
     "email-validator": "^2.0.4",
