@@ -1,25 +1,25 @@
-import knex = require("knex")
-import supertest = require("supertest")
-import { createLogger } from "winston"
-import { createApp } from "../src/app"
-import { Config } from "../src/config"
-import { Knex } from "../src/knex"
-import { Logger } from "../src/logger"
+import { createLogger } from 'winston'
+import { createApp } from '../src/app'
+import { Config } from '../src/config'
+import { Knex } from '../src/knex'
+import { Logger } from '../src/logger'
+import knex = require('knex')
+import supertest = require('supertest')
 
-const migrationsDir = `${__dirname}/../../../migrations`
-const seedsDir = `${__dirname}/../../../seeds`
+const migrationsDir = `${__dirname}/../migrations`
+const seedsDir = `${__dirname}/../seeds`
 
 const accessToken =
-  "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOjEsImlhdCI6MTUxNjIzOTAyMn0.Fknsf_nSFNdqS9JkFJABEEtMVffv9zR1_nrI2mAVx60"
+  'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOjEsImlhdCI6MTUxNjIzOTAyMn0.Fknsf_nSFNdqS9JkFJABEEtMVffv9zR1_nrI2mAVx60'
 
 const config = new Config({
-  AUTH_SERVER_TOKEN_SECRET: "secret",
-  AUTH_SERVER_DATABASE_URL: ":memory:",
-  AUTH_SERVER_DATABASE_CLIENT: "sqlite3",
-  AUTH_SERVER_ACCESS_TOKEN_LIFETIME: "30",
-  AUTH_SERVER_REFRESH_TOKEN_LIFETIME: "2592000",
-  AUTH_SERVER_ALLOWED_ORIGIN: "*",
-  PORT: "8080",
+  TOKEN_SECRET: 'secret',
+  DATABASE_URL: ':memory:',
+  DATABASE_CLIENT: 'sqlite3',
+  ACCESS_TOKEN_LIFETIME: '30',
+  REFRESH_TOKEN_LIFETIME: '2592000',
+  ALLOWED_ORIGIN: '*',
+  PORT: '8080',
 })
 
 let request: supertest.SuperTest<supertest.Test>
@@ -48,59 +48,48 @@ beforeEach(async () => {
   request = supertest(createApp(config, knexConnection, logger).callback())
 })
 
-describe("on POST /signup", () => {
-  test("should fail with 400 if no email specified", async () => {
+describe('on POST /signup', () => {
+  test('should fail with 400 if no email specified', async () => {
+    await request.post('/signup').send({ password: 'somepassword' }).expect(400)
+  })
+
+  test('should fail with 400 if email is not invalid', async () => {
     await request
-      .post("/signup")
-      .send({ password: "somepassword" })
+      .post('/signup')
+      .send({ email: 'invalid_email', password: 'somepassword' })
       .expect(400)
   })
 
-  test("should fail with 400 if email is not invalid", async () => {
-    await request
-      .post("/signup")
-      .send({ email: "invalid_email", password: "somepassword" })
-      .expect(400)
+  test('should fail with 400 if no password specified', async () => {
+    await request.post('/signup').send({ email: 'someone@mail.com' }).expect(400)
   })
 
-  test("should fail with 400 if no password specified", async () => {
-    await request
-      .post("/signup")
-      .send({ email: "someone@mail.com" })
-      .expect(400)
+  test('should fail with 400 if password is too short', async () => {
+    await request.post('/signup').send({ email: 'someone@mail.com', password: 'short' }).expect(400)
   })
 
-  test("should fail with 400 if password is too short", async () => {
-    await request
-      .post("/signup")
-      .send({ email: "someone@mail.com", password: "short" })
-      .expect(400)
+  test('should fail with 400 when request body is empty', async () => {
+    await request.post('/signup').expect(400)
   })
 
-  test("should fail with 400 when request body is empty", async () => {
-    await request.post("/signup").expect(400)
-  })
-
-  test("should fail with 409 if email already used by someone else", async () => {
+  test('should fail with 409 if email already used by someone else', async () => {
     await request
-      .post("/signup")
-      .send({ email: "foo@bar.baz", password: "somepassword" })
+      .post('/signup')
+      .send({ email: 'foo@bar.baz', password: 'somepassword' })
       .expect(409)
   })
 
-  test("should create user", async () => {
+  test('should create user', async () => {
     await request
-      .post("/signup")
-      .send({ email: "someone@mail.com", password: "somepassword" })
+      .post('/signup')
+      .send({ email: 'someone@mail.com', password: 'somepassword' })
       .expect(200)
 
     await expect(
-      knexConnection("users")
-        .where({ email: "someone@mail.com" })
-        .first(),
+      knexConnection('users').where({ email: 'someone@mail.com' }).first(),
     ).resolves.toEqual({
       id: expect.any(Number),
-      email: "someone@mail.com",
+      email: 'someone@mail.com',
       password: expect.any(String),
       created_at: expect.any(String),
       updated_at: expect.any(String),
@@ -108,9 +97,11 @@ describe("on POST /signup", () => {
   })
 })
 
-describe("/login", () => {
-  test("POST /login - should respond with authorization data", async () => {
-    const response = await request.post("/login").send({ email: "foo@bar.baz", password: "somepassword" })
+describe('/login', () => {
+  test('POST /login - should respond with authorization data', async () => {
+    const response = await request
+      .post('/login')
+      .send({ email: 'foo@bar.baz', password: 'somepassword' })
 
     expect(response.status).toBe(200)
     expect(response.body).toEqual({
@@ -118,11 +109,7 @@ describe("/login", () => {
       access_token: expect.any(String),
     })
 
-    await expect(
-      knexConnection("refresh_tokens")
-        .where({ user_id: 1 })
-        .first(),
-    ).resolves.toEqual({
+    await expect(knexConnection('refresh_tokens').where({ user_id: 1 }).first()).resolves.toEqual({
       id: expect.any(Number),
       user_id: 1,
       refresh_token: expect.any(String),
@@ -131,30 +118,22 @@ describe("/login", () => {
     })
   })
 
-  // eslint-disable-next-line jest/expect-expect
-  test("POST /login - should fail if email or password not specified", async () => {
-    await request.post("/login").expect(400)
+  test('POST /login - should fail if email or password not specified', async () => {
+    await request.post('/login').expect(400)
   })
 
-  // eslint-disable-next-line jest/expect-expect
   test("POST /login - should fail if email or password aren't correct", async () => {
-    await request
-      .post("/login")
-      .send({ email: "foo@bar.baz", password: "wrong" })
-      .expect(401)
+    await request.post('/login').send({ email: 'foo@bar.baz', password: 'wrong' }).expect(401)
 
-    await request
-      .post("/login")
-      .send({ email: "wrong@bar.baz", password: "wrong" })
-      .expect(401)
+    await request.post('/login').send({ email: 'wrong@bar.baz', password: 'wrong' }).expect(401)
   })
 })
 
-describe("/refreshToken", () => {
-  test("POST /refreshToken - should update refresh token and create new access token", async () => {
+describe('/refreshToken', () => {
+  test('POST /refreshToken - should update refresh token and create new access token', async () => {
     const response = await request
-      .post("/refreshToken")
-      .send({ refresh_token: "8e6112346a91d135e3cb8bbad7f5363eae2108ff" })
+      .post('/refreshToken')
+      .send({ refresh_token: '8e6112346a91d135e3cb8bbad7f5363eae2108ff' })
 
     expect(response.status).toBe(200)
     expect(response.body).toEqual({
@@ -163,81 +142,63 @@ describe("/refreshToken", () => {
     })
   })
 
-  test("POST /refreshToken - refresh token should be used only once", async () => {
+  test('POST /refreshToken - refresh token should be used only once', async () => {
     await request
-      .post("/refreshToken")
-      .send({ refresh_token: "8e6112346a91d135e3cb8bbad7f5363eae2108ff" })
+      .post('/refreshToken')
+      .send({ refresh_token: '8e6112346a91d135e3cb8bbad7f5363eae2108ff' })
       .expect(200)
 
     await request
-      .post("/refreshToken")
-      .send({ refresh_token: "8e6112346a91d135e3cb8bbad7f5363eae2108ff" })
+      .post('/refreshToken')
+      .send({ refresh_token: '8e6112346a91d135e3cb8bbad7f5363eae2108ff' })
       .expect(401)
   })
 
-  // eslint-disable-next-line jest/expect-expect
   test("POST /refreshToken - should fail if refresh token isn't valid", async () => {
-    await request
-      .post("/refreshToken")
-      .send({ refresh_token: "invalid token" })
-      .expect(401)
+    await request.post('/refreshToken').send({ refresh_token: 'invalid token' }).expect(401)
   })
 
-  // eslint-disable-next-line jest/expect-expect
   test("POST /refreshToken - should fail if refresh token isn't specified", async () => {
-    await request.post("/refreshToken").expect(400)
+    await request.post('/refreshToken').expect(400)
   })
 
-  // eslint-disable-next-line jest/expect-expect
-  test("POST /refreshToken - should fail if refresh token is expired", async () => {
+  test('POST /refreshToken - should fail if refresh token is expired', async () => {
     await request
-      .post("/refreshToken")
-      .send({ refresh_token: "dcb76e25b2079ee652d28f732f6679c441291d2e" })
+      .post('/refreshToken')
+      .send({ refresh_token: 'dcb76e25b2079ee652d28f732f6679c441291d2e' })
       .expect(401)
   })
 })
 
-describe("/me", () => {
-  // eslint-disable-next-line jest/expect-expect
-  test("GET /me - should get authorized user details", async () => {
-    await request
-      .get("/me")
-      .set("Authorization", `Bearer ${accessToken}`)
-      .expect(200, {
-        email: "foo@bar.baz",
-      })
+describe('/me', () => {
+  test('GET /me - should get authorized user details', async () => {
+    await request.get('/me').set('Authorization', `Bearer ${accessToken}`).expect(200, {
+      email: 'foo@bar.baz',
+    })
   })
 
-  // eslint-disable-next-line jest/expect-expect
-  test("GET /me - should fail if unauthorized", async () => {
-    await request.get("/me").expect(401)
+  test('GET /me - should fail if unauthorized', async () => {
+    await request.get('/me').expect(401)
   })
 
-  // eslint-disable-next-line jest/expect-expect
-  test("GET /me - should fail if authorization is wrong", async () => {
-    await request
-      .get("/me")
-      .set("Authorization", "Bearer Wrong")
-      .expect(401)
+  test('GET /me - should fail if authorization is wrong', async () => {
+    await request.get('/me').set('Authorization', 'Bearer Wrong').expect(401)
   })
 })
 
-describe("POST /forgotToken", () => {
-  it("should respond with 200 if success", async () => {
+describe('POST /forgotToken', () => {
+  it('should respond with 200 if success', async () => {
     await request
-      .post("/forgotToken")
-      .send({ refresh_token: "8e6112346a91d135e3cb8bbad7f5363eae2108ff" })
+      .post('/forgotToken')
+      .send({ refresh_token: '8e6112346a91d135e3cb8bbad7f5363eae2108ff' })
       .expect(200)
   })
 
-  it("should respond with 400 if no refresh token specified", async () => {
-    await request.post("/forgotToken").expect(400)
+  it('should respond with 400 if no refresh token specified', async () => {
+    await request.post('/forgotToken').expect(400)
   })
 
-  it("should respond with 401 if refresh token is not valid", async () => {
-    await request
-      .post("/forgotToken")
-      .send({ refresh_token: "invalid_refresh_token" })
-      .expect(401)
+  it('should respond with 401 if refresh token is not valid', async () => {
+    await request.post('/forgotToken').send({ refresh_token: 'invalid_refresh_token' }).expect(401)
   })
 })
